refactor upload route controller middleware
