+++ middlewares/uploadMiddleware/index.js
@@ -1,0 +1,5 @@
+const createPrivateUpload = require('./createPrivateUpload');
+const createPublicUpload = require('./createPublicUpload');
+const uploadToStorage = require('./uploadToStorage');
+
+module.exports = { uploadToStorage, createPrivateUpload, createPublicUpload };
