+++ routes/coreRoutes/coreApi.js
@@ -8,6 +8,12 @@ const router = express.Router();
 
 const adminController = require('@/controllers/coreControllers/adminController');
 const settingController = require('@/controllers/coreControllers/settingController');
+
+const {
+  createPublicUpload,
+  uploadToStorage,
+  createPrivateUpload,
+} = require('@/middlewares/uploadMiddleware');
 // //_______________________________ Admin management_______________________________
 
 var adminPhotoStorage = multer.diskStorage({
@@ -46,5 +52,37 @@ router.route('/setting/update/:id').patch(catchErrors(settingController.update))
 router.route('/setting/search').get(catchErrors(settingController.search));
 router.route('/setting/list').get(catchErrors(settingController.list));
 router.route('/setting/filter').get(catchErrors(settingController.filter));
+
+// //____________________________________________ API for Upload controller _________________
+
+router.route('/public/upload/:model/:fieldId').post(
+  uploadToStorage.array('upload', 100),
+  createPublicUpload,
+  // need to add proper controller
+  catchErrors((req, res) => {
+    if (req.upload.files) {
+      return res.status(200).send({
+        success: true,
+        result: req.upload.files,
+        message: 'File uploaded successfully!',
+      });
+    }
+  })
+);
+
+router.route('/private/upload/:model/:fieldId').post(
+  uploadToStorage.array('upload', 100),
+  createPrivateUpload,
+  // need to add proper controller
+  catchErrors((req, res) => {
+    if (req.upload.files) {
+      return res.status(200).send({
+        success: true,
+        result: req.upload.files,
+        message: 'File uploaded successfully!',
+      });
+    }
+  })
+);
 
 module.exports = router;
